backend/scripts/requirements.txt
torch>=1.9.0
torchvision>=0.10.0
timm>=0.4.12
pillow==10.0.0
numpy==1.24.3
